Move MapperContext to the public API
